// client/src/generated/Me.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { CharacterInfo } from "./CharacterInfo";

/**
 * The logged-in account and its characters.
 */
export type Me = { name: string, characters: Array<CharacterInfo>, max_characters: number, };